% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_parsers.R
\name{convert_to_phylo}
\alias{convert_to_phylo}
\title{Converting clustering object into phylo object}
\usage{
convert_to_phylo(cl.obj)
}
\arguments{
\item{cl.obj}{Clustering object}
}
\value{
Returns a phylo object
}
\description{
Converting clustering object into phylo object
}
