% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVL.R
\name{CVL}
\alias{CVL}
\title{Cross validation Loss for hierarchical clustering methods}
\usage{
CVL(
  data,
  cl.list,
  dists = "euclidean",
  mixed_dist = "gower",
  tol = 1e-20,
  seed = 99,
  median = F
)
}
\description{
Cross validation Loss for hierarchical clustering methods
}
