% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo_dendrogram.R
\name{evo_dendrogram}
\alias{evo_dendrogram}
\title{Plot evolution dendogram for given data and selected variables}
\usage{
evo_dendrogram(
  data,
  var_col,
  hclust_obj,
  tip_names = TRUE,
  scale_data = TRUE,
  fsize = c(0.9, 0.8),
  outline = FALSE,
  lwd = c(3, 7),
  leg_txt = NA,
  palette_discrete = "Set1",
  cex = c(0.4, 0.25),
  ...
)
}
\description{
Plot evolution dendogram for given data and selected variables
}
