% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_score.R
\name{L_score}
\alias{L_score}
\title{Leave one feature out cross validation loss}
\usage{
L_score(dend, original_data, seed = 99, ncores = 2, score = FALSE)
}
\arguments{
\item{dend}{Dendrogram of interest (an object of class "phylo")}

\item{original_data}{Entire dataset (data.frame or matrix) or feature of interest (numeric or factor vector) to compute cross validate loss}

\item{seed}{Fixed random seed for SIMMAP algorithm. Default is 99.}

\item{ncores}{Number of cores to paralelize the computing process. Default is 2.}

\item{score}{Set whether a vector of leave one feature out loss should be computed (TRUE) or compute the overall mean of the loss
directly (FALSE)}
}
\value{
A numeric vector of leave one feature out losses.
}
\description{
Leave one feature out cross validation loss for a given dendrogram and dataset. This is an auxiliary function to
the main functions CVL and PFIS but can be used separately.
}
