% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIS.R
\name{PFIS}
\alias{PFIS}
\title{Phylogenetical Feature Importance Score for Hierarchical clustering methods}
\usage{
PFIS(
  data,
  cl.list,
  dists = "euclidean",
  mixed_dist = "gower",
  tol = 1e-20,
  seed = 99,
  ncores = 2,
  plot = TRUE
)
}
\description{
Phylogenetical Feature Importance Score for Hierarchical clustering methods
}
