% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVL.R
\name{CVL}
\alias{CVL}
\title{Cross validation Loss for hierarchical clustering methods}
\usage{
CVL(
  data,
  cl.list,
  dists = "euclidean",
  mixed_dist = "gower",
  seed = 99,
  ncores = 2,
  median = FALSE
)
}
\arguments{
\item{data}{Data frame used to build hierarchical clustering}

\item{cl.list}{Named list of hierarchical clustering methods: each entry is composed by the name of the hierarchical
clustering algorithm (e.g. hclust) followed by a character vector of all it's underlying agglomeration methods selected
for comparison (e.g. "complete", "ward.D2", ...).}

\item{dists}{Chosen distance (or list of distances) to compute dissimilarity matrix for each hierarchical clustering method.
Default is "euclidean".}

\item{mixed_dist}{Chosen mixed distance (or list of mixed distances) to compute dissimilarity matrix for each hierarchical
clustering method. Mixed distances are preferable to use when there are categorical features present in the dataset.
Default is "gower".}

\item{seed}{Fixed random seed for SIMMAP algorithm. Default is 99.}

\item{ncores}{Number of cores to paralelize the computing process. Default is 2.}

\item{median}{Whether to compute the median of the leave one feature out loss associated to every feature or the mean.}
}
\value{
A data frame comparing the CVL of each hierarchical clustering method and distance combination.
}
\description{
Leave one feature out Cross validation loss for comparing a list of hierarchical clustering methods.
}
